css/css-view-transitions/parsing/view-transition-name-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS View Transitions: view-transition-name with computed values</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-view-transitions-1/">
<meta name="assert" content="view-transition-name supports custom keywords.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id=target></div>
<script>
test_computed_value("view-transition-name", "none");
test_computed_value("view-transition-name", "foo");
test_computed_value("view-transition-name", "bar");
test_computed_value("view-transition-name", "baz");
test_computed_value("view-transition-name", "unset", "none");
test_computed_value("view-transition-name", "initial", "none");
test_computed_value("view-transition-name", "inherit", "none");
test_computed_value("view-transition-name", "revert", "none");
test_computed_value("view-transition-name", "revert-layer", "none");

// It seems the results are different in different browsers:
// Gecko serializes this as "match-element",
// WebKit doesn't serialize this, i.e. empty string.
// However, it seems "match-element" makes more sense so we use it as the
// result. Also, auto keyword is still under discussion, so we don't include it
// for now. See https://github.com/w3c/csswg-drafts/issues/12091 for more
// information.
test_computed_value("view-transition-name", "match-element");
test_computed_value("view-transition-name", "maTch-element", "match-element");
</script>
</body>
</html>
